requirements.txt
exceptiongroup==1.2.1
iniconfig==2.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tomli==2.0.1
tzdata==2024.1
